specialists/agents/requirements.txt
fastapi==0.116.0
uvicorn==0.22.0
pydantic==2.11.7
pydantic-settings==2.10.1
google-generativeai==0.3.1